fix: s3 upload client, restructure tests
